tests: split utils into multiple units
* also rename ‹test_config› to just ‹config› since pytest doesn't care
  and it's just redundant
